+++ ksql-common/src/main/java/io/confluent/ksql/metrics/ConsumerCollector.java
@@ -140,14 +140,9 @@ public class ConsumerCollector implements MetricCollector {
 
 
   @Override
-  public double currentMessageProductionRate() {
-    return 0;
-  }
-
-  @Override
   public double currentMessageConsumptionRate() {
     final List<TopicSensors.Stat> allStats = new ArrayList<>();
-    topicSensors.values().stream().forEach(record -> allStats.addAll(record.stats(false)));
+    topicSensors.values().forEach(record -> allStats.addAll(record.stats(false)));
 
     return allStats
         .stream()
