+++ ksql-common/src/main/java/io/confluent/ksql/metrics/ProducerCollector.java
@@ -129,14 +129,9 @@ public class ProducerCollector implements MetricCollector {
   }
 
   @Override
-  public double currentMessageConsumptionRate() {
-    return 0;
-  }
-
-  @Override
   public double currentMessageProductionRate() {
     final List<TopicSensors.Stat> allStats = new ArrayList<>();
-    topicSensors.values().stream().forEach(record -> allStats.addAll(record.stats(false)));
+    topicSensors.values().forEach(record -> allStats.addAll(record.stats(false)));
 
     return allStats
         .stream()
